fix process interrupt on Windows
on *NIX, we can simply send SIGINT to the GDB process, but on Windows we need to use `GenerateConsoleCtrlEvent` with `CTRL_C_EVENT` to interrupt the GDB process without terminating the frontend.

But on Windows, doing so would also terminate the parent process (the frontend) if we don't install a custom `SetConsoleCtrlHandler` to ignore the `CTRL_C_EVENT` and `CTRL_BREAK_EVENT` for our process.